// Hydra-FE/app/features/products/types/EnhancedCard.types.ts
export interface CardData {
  id: string;
  title: string;
  subtitle?: string;
  price: string;
  imageUrl: string;
  grade?: string;
  gradeColor?: string;
  href?: string;
  stock?: number;
  expansion?: string;
  variant?: string;
  cardName?: string;
  condition?: string;
  language?: string;
  immediateDelivery?: boolean;
  isLocalInventory?: boolean;
  foil?: boolean;
  surgeFoil?: boolean;
  cardNumber?: string;
  metadata?: string[];
  images?: string[];
  tags?: string[];
  importationId?: string | null;
  originalPrice?: string;
  rating?: number;
  reviewCount?: number;
  isNew?: boolean;
  isOnSale?: boolean;
  discountPercentage?: number;
  isBundle?: boolean;
  category?: string;
  tcg?: string;
  tcgId?: string;
  price_mxn_local?: number;
  price_mxn_importation?: number;
  soldBy?: string;
  storeLogo?: string | null;
}